pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-llm-chat"
version = "0.1.0"
description = "A secure interactive REPL with GPT-4 and rate limiting."
authors = [{name="Sean Rawlins", email="srawlins@gmail.com"}]
dependencies = [
    "openai",
    "psycopg2-binary",
    "bcrypt",
    "pyjwt",
    "redis",
    "sqlalchemy",
    "alembic",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "email-validator",
    "authlib",
    "requests",
    "python-multipart"
]
requires-python = ">=3.8"

[project.scripts]
mini-llm-chat = "mini_llm_chat.cli:main"
mini-llm-chat-web = "mini_llm_chat.web:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "mini_llm_chat/tests/.*",
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".eggs",
    "*.egg",
    "build",
    "dist",
]

[tool.pytest.ini_options]
testpaths = ["mini_llm_chat/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
